fixup! feat: add external schema refs and minimal defs
