--- internal/opendpi/info.go
@@ -1,11 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright 2026 Daco Labs
-
-package opendpi
-
-// Info contains metadata about the data product.
-type Info struct {
-	Title       string
-	Version     string
-	Description string
-}
