--- internal/opendpi/tag.go
@@ -1,10 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// Copyright 2026 Daco Labs
-
-package opendpi
-
-// Tag is used for categorizing ports.
-type Tag struct {
-	Name        string
-	Description string
-}
